+++ CHANGELOG.md
@@ -1,6 +1,27 @@
 # Changelog
 All notable changes to this package will be documented in this file.
 
+## [1.15.17] - 2022-04-27
+
+Plastic SCM for Unity is now available as part of the Version Control Package! You can enable Plastic SCM via Window > Plastic SCM to get started!
+If you have previously used the Unity Asset Store Plastic SCM plug-in, you can now simply use this package. Make sure you delete the plug-in from your project.
+Removing a previously added Plastic SCM Asset Store Plug-In:
+- Select the PlasticSCM folder in the Assets\Plugins folder on the Project tab, then click Edit > Delete
+- Close the Unity Editor and open your project again. You will find the Plastic SCM menu item in the Window menu.
+
+### Added
+- Added checkin comment column to Incoming Changes view
+
+### Changed
+- Updated Go Back confirmation message to be consistent with feature
+- Updated Create Child Branch dialog to focus on branch name field when opened
+- Improved messaging of Subtractive Merge after using Go Back feature
+
+### Fixed
+- Fixed assets not added correctly when Plastic SCM window is not open
+- Fixed wrong position of overlay icons on Pending Changes view
+- Disallowed Go Back feature to a changeset from another branch
+
 ## [1.15.16] - 2022-03-28
 
 Plastic SCM for Unity is now available as part of the Version Control Package! You can enable Plastic SCM via Window > Plastic SCM to get started!